Referral code issue fixed Wallet amount issue fixed

+++ app/src/main/java/com/oyola/app/models/Invoice.java
@@ -26,7 +26,7 @@ public class Invoice {
     private double discount;
     @SerializedName("promocode_amount")
     @Expose
-    private int promocode_amount;
+    private double promocode_amount;
     @SerializedName("delivery_charge")
     @Expose
     private double deliveryCharge;
@@ -106,11 +106,11 @@ public class Invoice {
         this.discount = discount;
     }
 
-    public int getPromocode_amount() {
+    public double getPromocode_amount() {
         return promocode_amount;
     }
 
-    public void setPromocode_amount(int promocode_amount) {
+    public void setPromocode_amount(double promocode_amount) {
         this.promocode_amount = promocode_amount;
     }
